minor refactoring, adding file headers
