--- titus-common-ext/elasticsearch/src/main/java/com/netflix/titus/ext/elasticsearch/IndexHeader.java
@@ -1,32 +1,0 @@
-package com.netflix.titus.ext.elasticsearch;
-
-public class IndexHeader {
-    private String _index;
-    private String _type;
-    private String _id;
-
-    public String get_index() {
-        return _index;
-    }
-
-    public void set_index(String _index) {
-        this._index = _index;
-    }
-
-    public String get_type() {
-        return _type;
-    }
-
-    public void set_type(String _type) {
-        this._type = _type;
-    }
-
-    public String get_id() {
-        return _id;
-    }
-
-    public void set_id(String _id) {
-        this._id = _id;
-    }
-}
-
