+++ titus-common-ext/elasticsearch/src/test/java/com/netflix/titus/ext/elasticsearch/DefaultEsClientIntegrationTest.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2020 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.netflix.titus.ext.elasticsearch;
 
 import java.time.Instant;
@@ -18,11 +33,10 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 public class DefaultEsClientIntegrationTest {
     @ClassRule
     public static final EsExternalResource ES_RESOURCE = new EsExternalResource();
+    private static final ParameterizedTypeReference<EsRespSrc<TestDoc>> esRespTypeRef = new ParameterizedTypeReference<EsRespSrc<TestDoc>>() {
+    };
 
     private DefaultEsClient<TestDoc> client;
-
-    private static ParameterizedTypeReference<EsRespSrc<TestDoc>> esRespTypeRef = new ParameterizedTypeReference<EsRespSrc<TestDoc>>() {
-    };
 
     @Before
     public void setup() {
